requirements.txt
# Core dependencies with exact versions
aiogram==3.14.0
fastapi==0.115.5
uvicorn[standard]==0.32.1
asyncpg==0.30.0
python-dotenv==1.0.0

# Data validation
pydantic==2.9.2
pydantic-settings==2.6.1

# HTTP and networking
aiohttp==3.10.10
httpx==0.25.0

# Scheduler
apscheduler==3.10.4

# Logging
structlog==24.4.0
python-json-logger==2.0.7

# Web framework extras
python-multipart==0.0.17
jinja2==3.1.4

# Security
bleach==6.1.0